Forbid interchanging label refs and integers
A label reference used in place of an integer argument is resolved as
usual, but must have a corresponding definition, so is not useful for
automatically assigned variables. An integer used in place of a label
reference argument references a label by its generated number, which is
unstable. Label definitions cannot use numbers.

Separate label refs and integers in the grammar to simplify this.